Initial Dump commit for @wrigby
This is just a full solution dump of my current Cryptopals
challenges work.  It is not nicely divided by Sets because
the code doesn't divide nicely that way.

Most code is in the AESTools as that is the bulk of the
work, though XORTools and KeyUtils also have some useful
code.